requirements.txt
fastapi>=0.109.1
uvicorn>=0.27.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
pymongo>=4.6.1
motor>=3.3.2
python-multipart>=0.0.9
email-validator>=2.1.0
phonenumbers>=8.13.30
python-magic>=0.4.27
pandas>=2.2.0
PyYAML>=6.0.1